refactor: rename commands.js to questionHandler.js
